requirements.txt
appdirs==1.4.0
asgiref==1.0.0
attrs==16.3.0
autobahn==0.17.1
Automat==0.5.0
Babel==2.3.4
channels==1.0.3
constantly==15.1.0
daphne==1.0.3
decorator==4.0.11
Django==1.10.5
django-authtools==1.5.0
django-extensions==1.7.6
django-multiselectfield==0.1.5
django-phonenumber-field==1.1.0
fancycompleter==0.7
incremental==16.10.1
ipython==5.2.2
ipython-genutils==0.1.0
ordereddict==1.1
packaging==16.8
pdbpp==0.8.3
pexpect==4.2.1
phonenumberslite==8.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
pytz==2016.10
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
Twisted==17.1.0
txaio==2.6.1
wcwidth==0.1.7
wmctrl==0.3
zope.interface==4.3.3
